requirements.txt
idna==2.6
requests==2.18.4
cryptography==2.3.1
Flask-Cors==3.0.3
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.3.2
opcua==0.98.5
psycopg2-binary==2.7.5
SQLAlchemy==1.1.14
PyYaml==3.13
simplejson==3.11.1